+++ changelog.md
@@ -11,6 +11,7 @@
 * Modbus read methods using libmodbus api added
 * Keeping a dummy variable in holding register method for testing
 * Auto detection of modbus data type based on register address and exit if address out of range
+* Tested with modbus slave simulator [OpenModSim](https://github.com/sanny32/OpenModSim)
 
 **Changed** 
 * Config data structure changed for better parsing